requirements.txt
flask 
flask-sqlalchemy 
flask-login 
flask-wtf 
flask-migrate 
bootstrap-flask
email_validator
